0.2.0 Changed middleware to use req.dbClient
I realised that my architecture was wrong for how I wanted to use the pg
package. While I wanted to pass a pg client in to the middleware, I
didn't want a single client but a client passed in as part of the
request. As such, I intend to have a further middleware that generates a
client